Added boundguard for /randomelody
So, optionally, you can ensure that all tones are playable on the
keyboard.

Once the tones go beyond the keys to the right-hand side, they are
moved down by an octave or more, depending on how far they go.

Once they go beyond the keys to the left-hand-side, they are moved down
by an octave or more, depending on how far they go.

So, they are kind of band-limited.