single-sompyler-procman.sh: Switched gt -> ge (date_of)
gt does not support execution that is too fast. ge would not work either
in environments with high-speed network and high-speed bots or so, i.e.
once evaluation of audiofile results in the score being fixed and
uploaded, we assume it would be at least a second older than the score
file.